requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0

